adapted models, separated interfaces from classes
