+++ frontend/src/app/_models/studysmarter/index.ts
@@ -1,5 +1,7 @@
 export * from './flashcard.class';
+export * from './flashcard.interface';
 export * from './login-request.interface';
 export * from './login-response.interface';
 export * from './response.interface';
+export * from './subject.class';
 export * from './subject.interface';
